Reconcile Phase 2 with the receipts memoization that landed on main
The rebase onto main brings in #732, which rewrote the same function this
phase's deliverable 1 rewrites. Both survive; neither was reverted to make the
merge easy.

CONTENT IS A Cow, NOT AN OWNED String OR A BARE BORROW
#732 made `BlockDraft.content` a borrowed `&'a str` for a specific reason: a
block's bytes ride `BlockRegistered`, which fires once per block per turn, so
owning them meant re-cloning the system prefix — the largest gap block there is
— on every step for the rest of the turn. Deliverable 1 splits recall into
per-frame blocks, whose text is a borrowed slice of the same message and needs
no allocation, but it also decomposes attachments, whose preimage is a freshly
serialized JSON string that nothing owns.

`Option<Cow<'a, str>>` is the resolution that concedes nothing: the prefix and
every recall segment stay borrowed, and only the one site that genuinely mints
bytes allocates. Reverting to `Option<String>` would have merged cleanly and
silently undone a merged performance fix.

ATTACHMENTS GO THROUGH THE MEMO, NOT AROUND IT
Deliverable 1 built attachment blocks with a `BlockDraft::new` constructor.
#732 removed that constructor — construction is inline, through a `push`
closure that consults `BlockDigestCache`. Attachments now route through `push`
like every other kind, so the memo covers them too; a hydrated attachment
passes `None` for the local preimage, which keeps its inline base64 out of the
event stream while still hashing the real value for identity.

The two rewrites turned out to be complementary rather than competing:
splitting a message into more blocks makes the position-keyed memo finer, never
absent, so a decomposed recall block is still hashed once per turn.

RECEIPTS.RS IS SPLIT
The merged file reached 1532 lines, over the 1500 ratchet, and a baseline entry
is not allowed for a file that crosses it. The inline `mod tests` moves to a
sibling `receipts/tests.rs` — a pure move, no assertion touched — taking the
module to 949.

The signal that matters is green: both byte-exact reconstruction tests pass
with recall now decomposed —
`reconstruct::tests::reconstructs_a_tool_round_trip_byte_exact_and_verified_from_the_fold`
and
`agent::persistence::stream_tests::end_to_end_receipt_reconstructs_the_step_byte_exact_from_the_persisted_store`.

`make gate` green: exit 0, 3698 tests, check-file-size OK.
